Add delegates to atom properties table
