+++ MDANSE_GUI/Src/MDANSE_GUI/Tabs/Views/PlotDetailsView.py
@@ -18,61 +18,8 @@ from qtpy.QtWidgets import QTreeView, QComboBox, QItemDelegate, QColorDialog
 from qtpy.QtCore import Signal, Slot, Qt
 from qtpy.QtGui import QColor
 
+from MDANSE_GUI.Tabs.Views.Delegates import ColourPicker, MplStyleCombo
 from MDANSE_GUI.Tabs.Models.PlottingContext import get_mpl_lines, get_mpl_markers
-
-
-class ColourPicker(QItemDelegate):
-
-    def __init__(self, *args, **kwargs) -> None:
-        super().__init__(*args, **kwargs)
-
-    def createEditor(self, parent, option, index):
-        dialog = QColorDialog(parent)
-        return dialog
-
-    def setEditorData(self, editor, index):
-        colour_string = index.data()
-        color = QColor(colour_string)
-        editor.setCurrentColor(color)
-
-    def setModelData(self, editor, model, index):
-        if editor.result() == QColorDialog.DialogCode.Accepted:
-            color = editor.currentColor()
-            colour_string = color.toRgb()
-            model.setData(index, colour_string)
-
-
-class MplStyleCombo(QItemDelegate):
-
-    def __init__(self, *args, mpl_items=None, **kwargs) -> None:
-        self._items = list([str(x) for x in mpl_items.keys()])
-        super().__init__(*args, **kwargs)
-
-    def setEditorData(self, editor, index):
-        editor.blockSignals(True)
-        text = index.model().data(index, Qt.DisplayRole)
-        try:
-            i = self._items.index(text)
-        except ValueError:
-            i = 0
-        editor.setCurrentIndex(i)
-        editor.blockSignals(False)
-
-    def setModelData(self, editor, model, index):
-        model.setData(index, editor.currentText())
-
-    def createEditor(self, parent, option, index):
-        combo = QComboBox(parent)
-        li = []
-        for item in self._items:
-            li.append(item)
-        combo.addItems(li)
-        combo.currentIndexChanged.connect(self.currentIndexChanged)
-        return combo
-
-    @Slot()
-    def currentIndexChanged(self):
-        self.commitData.emit(self.sender())
 
 
 class PlotDetailsView(QTreeView):
